+++ changelog.md
@@ -3,6 +3,7 @@ v0.0.12-alpha
 * Allow clearing wand spells using the Wand Inscription Table
 * Add Auto-Charger, to automatically charge wands from nearby fonts
 * Philosophy research for Earth and Sky now requires the player to be in the Overworld
+* Add Dream Vision Talisman, a means to convert experience into observations
 
 v0.0.11-alpha
 -------------